# Conflicts:
#	BookService/src/main/java/com/cts/controller/BookController.java
#	BookService/src/main/java/com/cts/controller/CategoryController.java

+++ PaymentService/src/main/java/com/cts/service/PaymentServiceImpl.java
@@ -124,7 +124,7 @@ public class PaymentServiceImpl implements PaymentService {
 	private Long generateUniquePaymentId() {
 		Long newId;
 		do {
-			newId = UUID.randomUUID().getMostSignificantBits() & Long.MAX_VALUE;
+			newId = UUID.randomUUID().getMostSignificantBits() & Integer.MAX_VALUE;
 		} while (paymentRepository.existsByPaymentId(newId));
 
 		return newId;
